Common path for SQLite database
