+++ wikipedia/reader.py
@@ -5,7 +5,7 @@ Written by Michael Wheeler and Jay Sherman.
 """
 import bz2
 from argparse import ArgumentParser
-from config import WIKIPEDIA_ARCHIVE_FILE, WIKIPEDIA_INDEX_FILE, OUTPUT_DATA_DIR
+from config import WIKIPEDIA_ARCHIVE_FILE, WIKIPEDIA_INDEX_FILE, OUTPUT_DATA_DIR, SQLITE_ARCHIVE_INDEX_FILE
 from os.path import exists
 from os import PathLike
 import sqlite3
@@ -40,7 +40,7 @@ class WikipediaArchiveSearcher:
         Maps each known article title to its start index, end index, title, and unique ID for fast searching later.
         :return: connection to xml_indices database, which has table named articles that holds above info
         """
-        conn = sqlite3.connect(OUTPUT_DATA_DIR / "pages.db")
+        conn = sqlite3.connect(SQLITE_ARCHIVE_INDEX_FILE)
         return conn
 
     def retrieve_article_xml(self, title: WikipediaArticle) -> str:
